Move deployment.py to verta/deployment/ (#1085)
