+++ client/verta/verta/deployment/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from .deployedmodel import (
+    DeployedModel,
+    prediction_input_unpack,
+    prediction_io_cleanup,
+)
